Large commit message because I didn't understand a lot/want to capture what i learned:
Basically, i had to write two files, one was the server and one was the client, once i ran the server in file one, it was running on a continuous loop. then when i ran file two, the client file, it was able to querry the server and get the borring information from it (time/date and a message that says, closing connection, bye). but it was super interesting how it worked. next i'm working on using a server from the big bad internet, not my safe humble computer